server/requirements.txt
Flask==2.0.2
flask_cors==3.0.10
flask_pymongo==2.3.0
python-dotenv==0.19.2
